feat(v2): V2 适配壳 + 双格式入口（V1/V2 统一——对齐 visual-cache 模式）
