+++ package.json
@@ -10,7 +10,7 @@
       "types": "./dist/server.d.ts"
     },
     "./tui": {
-      "import": "./dist/tui.js",
+      "import": "./tui/index.js",
       "config": {
         "enabled": true
       }
@@ -19,6 +19,7 @@
   "files": [
     "dist",
     "src",
+    "tui",
     "install.mjs",
     "README.md"
   ],
